requirements.txt
kafka-python==2.0.1
azure-eventhub==1.3.*
azure-eventhub-checkpointstoreblob-aio==1.1.0
